requirements.txt
black==22.10.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.1.3
flake8==6.0.0
future==0.18.2
idna==3.4
isort==5.10.1
mccabe==0.7.0
mypy==0.991
mypy-extensions==0.4.3
oauthlib==3.2.2
pathspec==0.10.2
platformdirs==2.5.4
pycodestyle==2.10.0
pyflakes==3.0.1
python-twitter==3.5
requests==2.28.1
requests-oauthlib==1.3.1
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.13
